--- src/main/java/Duke/DukeException.java
@@ -1,7 +1,0 @@
-package Duke;
-
-public class DukeException extends Exception{
-    public DukeException(String message) {
-        super(message);
-    }
-}
