+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,10 @@
+package duke;
+
+/**
+ * Duke exception class which encapsulates all the errors.
+ */
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
