+++ src/main/java/duke/commands/DeadlineCommand.java
@@ -1,28 +1,30 @@
-package Duke.Commands;
-
-import Duke.DukeException;
-import Duke.Storage;
-import Duke.TaskList;
-import Duke.Tasks.Deadlines;
-import Duke.Tasks.ToDos;
-import Duke.Ui;
+package duke.commands;
 
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
 
-public class DeadlineCommand extends Command{
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.tasks.Deadlines;
+
+/**
+ * A command to create a deadline task.
+ */
+public class DeadlineCommand extends Command {
     protected TaskList taskList;
     protected Storage storage;
     protected Ui ui;
     protected String input;
 
     /**
-     * Constructor for a deadline command
+     * Deadline command constructor to create a deadline task.
      *
-     * @param input The input for the command
-     * @param taskList The taskList to store the task
-     * @param storage The storage system to store the list
-     * @param ui The ui to print the commands
+     * @param input The input for the command.
+     * @param taskList The taskList to store the task.
+     * @param storage The storage system to store the list.
+     * @param ui The ui to print the commands.
      */
     public DeadlineCommand(String input, TaskList taskList, Storage storage, Ui ui) {
         this.input = input;
@@ -32,9 +34,9 @@ public class DeadlineCommand extends Command{
     }
 
     /**
-     * Execute the command itself
+     * Execute the command itself.
      *
-     * @throws DukeException When there is an error saving
+     * @throws DukeException When there is an error saving.
      */
     @Override
     public void execute() throws DukeException {
