+++ src/main/java/duke/tasks/ToDos.java
@@ -1,9 +1,12 @@
-package Duke.Tasks;
+package duke.tasks;
 
+/**
+ * Represents a ToDo task.
+ */
 public class ToDos extends Task {
     /**
-     * Constructor for todo task
-     * @param description Description of task
+     * Constructor for todo task.
+     * @param description Description of task.
      */
     public ToDos(String description) {
         super(description);
